chore: SubscriptionManagement 페이지에서 KidInformationBox를 다른 파일로 분리
