+++ FE/src/pages/Driver/SubscriptionManagement/KidInformationBox/KidInformationBox.module.scss
@@ -1,0 +1,103 @@
+@import "@/styles/fontStyle.scss";
+@import "@/styles/color.scss";
+
+.content {
+    display: flex;
+    flex-direction: column;
+    gap: 30px;
+    writing-mode: horizontal-tb;
+    width: 90%;
+    margin: 10px auto;
+    padding: 20px;
+    border: 1px solid #c0c1c2;
+    border-radius: 10px;
+}
+
+.kidInfo {
+    display: flex;
+    flex-direction: row;
+    gap: 20px;
+    align-items: center;
+    justify-content: space-evenly;
+}
+
+.infoBox {
+    display: flex;
+    flex-direction: column;
+    // justify-content: space-between;
+    text-align: left;
+    gap: 10px;
+    width: 80%;
+
+    .name {
+        @extend %B_18;
+    }
+    .birthDay {
+        @extend %R_14;
+        color: #656363;
+    }
+}
+
+.kidImg {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    background-color: #c0c1c2;
+}
+
+.profiles {
+    display: flex;
+    align-items: center;
+}
+
+.status {
+    padding: 3px;
+    border-radius: 10px;
+    color: white;
+    @extend %R_12;
+}
+
+.accept {
+    background-color: $primary;
+}
+
+.waiting {
+    background-color: $blue;
+}
+
+.headMessage {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 450px;
+    @extend %B_18;
+}
+
+span {
+    @extend %R_14;
+    margin-right: 10px;
+}
+
+.btnBox {
+    display: flex;
+    gap: 20px;
+    justify-content: center;
+}
+
+.denyButton {
+    background-color: #ca6e6e;
+    width: 40%;
+    height: 50px;
+    border-radius: 10px;
+    color: white;
+    @extend %M_20;
+}
+
+.acceptButton {
+    background-color: #6eca95;
+    width: 40%;
+    height: 50px;
+    border-radius: 10px;
+    color: white;
+    @extend %M_20;
+}
